Email subscription notification on bug report state change
Call SubscriptionHelper.NotifyBugReportStateChanged in BugReportController when posting edited bug reports. Add GetAllSubscribed to IProjectRepository and implement in DapperProjectRepository - Get IEnumerable<int> of user IDs subscribed to the specific bug report ID. Add NotifyBugReportStateChanged in ISubscriptionHelper and implement in SubscriptionHelper - Handles generating the email to send notifying of a change to the bug report state.