Implement basic search results.
Result action method of SearchController takes the SearchResultsViewModel containing the text entered into the search box, or the results of a previous advanced search. The current project ID is determined, a collection of bug reports is accessed from the database, and then the reports are queried for the search term.