Implement XDSoft datepicker for advanced search
Results.cshtml references datetimepicker libraries. Add input boxes corresponding to DateRangeBegin and DateRangeEnd in the SearchExpression model, to handle selecting the start and end dates for the search term. Modify type of DateRangeBegin and DateRangeEnd to DateTime and set default date of DateRangeEnd to Now. In _Layout, move javascript sources infront of @RenderBody so that jQuery is declared first.