Starting date range defaults to project creation date
Add private GetProjectCreationTime method in SearchController which takes the project id, gets the project from the repository, and then returns the creation time. Get this DateTime and check whether the user has entered a more recent date into the date picker. If this is not the case then use the project creation date as default.