Add AddUserToProjectMemberRoleForAllPendingInvitations class to ProjectInvitation interface, implement, and call on registering user
Add AddUserToProjectMemberRoleForAllPendingInvitations method to IProjectInvitation interface, and implement in ProjectInvitation class - gets a list of project IDs from the repository which the user is pending invitations for. The user is then added to the 'Member' role via the user manager. This method is then called from the Register action method inside AccountController. Whenever a user is registered, that email address is checked for pending invitations.