Implement adding new milestone
Add New GET and POST action methods to MilestoneController - GET method adds project ID to the milestone object and passes to New view. POST method sends the milestone to AddMilestone repository method to add to the database. Add New view with form for entering the Title, Description, and DueDate properties for milestone creation.